메인페이지 디자인 수정
