+++ src/components/MainLayout/style.js
@@ -3,5 +3,5 @@ import { css } from "@emotion/react";
 export const layout = css`
   box-sizing: border-box;
   width: 100%;
-  height: 80px;
+  height: 100px;
 `;
