+++ src/components/MainMenu/style.js
@@ -5,31 +5,32 @@ export const container = css`
     display: flex;
     flex-direction: column;
     margin-top: 40px;
-    width: 100%;
+    width: 1200px;
     height: 100%;
 `;
 
 export const headerLayout = css`
     box-sizing: border-box;
     width: 100%;
-    border-bottom: 2px solid #000000;
 `;
 
 export const layout = css`
     box-sizing: border-box;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
     width: 100%;
 `;
 
 export const buttonLayout = css`
     box-sizing: border-box;
-    width: 100%;
     display: flex;
-    justify-content: center;
+    padding-left: 4px;
+    width: 100%;
 
     & a {
         box-sizing: border-box;
         margin-right: 60px;
-        margin-bottom: 5px;
         text-decoration: none;
         font-size: 16px;
         color: #5B5B5B;
@@ -47,16 +48,13 @@ export const menuLayout = css`
     display: flex;
     justify-content: center;
     align-items: center;
-    width: 1300px;
+    width: 1200px;
     height: 100%;
 `;
 
 export const menuListBox = css`
     box-sizing: border-box;
-    display: flex;
-    justify-content: center;
-    margin-top: 15px;
-    padding: 0px 50px 0px 50px;
+    margin-top: 20px;
     width: 100%;
     height: 100%;
 `;
